requirements.txt
docling~=2.28.0
PyYAML~=6.0.2

pytest~=8.3.5
setuptools~=68.2.0